requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
fake-useragent==0.1.11
free-proxy==1.0.5
h11==0.12.0
httmock==1.4.0
http-request-randomizer==1.3.2
idna==3.3
lxml==4.7.1
outcome==1.1.0
psutil==5.9.0
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
